pyproject.toml
[tool.poetry]
name = "notes"
version = "0.1.0"
description = ""
authors = ["yoshino"]
readme = "README.md"
packages = [ { include = "parse_question_pdf", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.0.9"
pypdf = "^3.17.1"
openai = "^1.3.7"
python-dotenv = "^1.0.0"
pdfminer-six = "^20221105"
pandas = "^2.1.3"
transformers = "^4.35.2"
torchvision = "^0.16.1"
torch = "^2.1.1"
matplotlib = "^3.8.2"
pdf2image = "^1.16.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
